Add fm-streamer rust binary
The fm-streamer binary is based on the simple_fm example [1].

It is intended to receive input from an rtl-sdr device, demodulate based
on command-line arguments and stream the demodulated data to a network
stream.

The current implementation listens to a hardcoded frequency, demodulates
and outputs a sound-stream to stdout. Can be tested using the following
command:

cargo run | play -r 32k -t raw -e s -b 16 -c 1 -V1 -

[1] https://github.com/ccostes/rtl-sdr-rs/blob/main/examples/simple_fm.rs

Signed-off-by: Fredrik Lönnegren <fredrik@frelon.se>